gcc compatibility tweaks : using -std=c99 flag when compiling with gcc on Linux to enable support for 'restrict' keyword. An alternative would be a more complex preprocessor macro, but gcc isn't evaluated by 42 anyway. libft now compiles with gcc

+++ ft_printf/sources/ftpf_lenmods_xp.c
@@ -43,7 +43,7 @@ void	ftpf_manage_lenmod_xp_sub(va_list *ap, t_ftpf_env *e, char **itoa_ret)
 {
 	if (ft_strchr("xX", e->conversion) && !e->mod[0])
 	{
-		if (!(e->param->u = (unsigned int)va_arg(*ap, unsigned int*)))
+		if (!(e->param->u = (unsigned long)va_arg(*ap, unsigned long*)))
 			e->isnull = 1;
 		*itoa_ret = ft_itoa_ull(e->param->u, 16);
 	}
